backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finguard-backend"
version = "0.1.0"
description = "FastAPI backend for FinGuard"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.1",
  "fastapi>=0.115.6",
  "httpx>=0.28.1",
  "pydantic-settings>=2.7.1",
  "pypdf>=5.4.0",
  "psycopg[binary]>=3.2.3",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
  "ruff>=0.9.1"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
